fix(core): add workshop story
